Port to micro:bit v2.21 nRF52833 board.
This commit adds some scaffolding for micro:bit v2.21 in the form of
new chip, board, and app folders. Most of the work is based on the
initial sections of Hubris on PineTime blog post
(https://artemis.sh/2022/03/28/oxide-hubris-on-pinetime.html), but
some reading of nRF52833 data sheet and hacking with `probe-rs` was
still necessary.

I've only added `jefe`, `jiffy`, and `idle` tasks, which is the bare
minimum (again, as per Hubris on PineTime blog post).